fix: type hint of load memories

+++ utils/memory.py
@@ -543,7 +543,7 @@ def recur_improv(txt_profile: str, txt_audio: str, txt_whisp_prompt: str, txt_ch
 
 @optional_typecheck
 @lru_cache
-def cached_load_memories(path: str, modtime: int) -> List[dict]:
+def cached_load_memories(path: str, modtime: float) -> List[dict]:
     "actual code that load the memories, but cached"
     path = Path(path)
     assert path.exists(), f"File not found: {path}"
